TrueNAS 12 to SCALE update
(cherry picked from commit 594154ba853d72de7ba2d787953408f4c83425fe)
