+++ src/middlewared/middlewared/plugins/update_/install_freebsd.py
@@ -1,4 +1,9 @@
 # -*- coding=utf-8 -*-
+import contextlib
+import os
+import shutil
+import subprocess
+
 from freenasOS import Configuration, Manifest, Update
 
 from middlewared.service import CallError, job, private, Service
@@ -7,10 +12,15 @@ from middlewared.worker import FakeJob
 from .utils import can_update
 from .utils_freebsd import UpdateHandler
 
+run_kw = dict(check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, encoding='utf-8', errors='ignore')
+
 
 class UpdateService(Service):
     @private
     def install_impl(self, job, location):
+        if os.path.exists(os.path.join(location, 'scale')):
+            return self._install_scale(job, os.path.join(location, 'update.sqsh'))
+
         old_manifest = Configuration.Configuration().SystemManifest()
 
         new_manifest = Manifest.Manifest(require_signature=True)
@@ -37,9 +47,33 @@ class UpdateService(Service):
 
     @private
     def install_manual_impl(self, job, path, dest_extracted):
+        p = subprocess.run(['file', path], stdout=subprocess.PIPE, stderr=subprocess.DEVNULL, encoding='ascii',
+                           errors='ignore')
+        if 'Squashfs filesystem' in p.stdout:
+            return self._install_scale(job, path)
+
         job.set_progress(30, 'Extracting file')
         Update.ExtractFrozenUpdate(path, dest_extracted, verbose=True)
 
         job.set_progress(50, 'Applying update')
         if self.install_impl(job, dest_extracted) is None:
             raise CallError('Uploaded file is not a manual update file')
+
+    def _install_scale(self, job, path):
+        location = os.path.dirname(path)
+
+        mounted = os.path.join(location, 'squashfs-root')
+        with contextlib.suppress(FileNotFoundError):
+            shutil.rmtree(mounted)
+
+        job.set_progress(0, 'Extracting update file')
+        subprocess.run(['unsquashfs', os.path.basename(path)], cwd=location, **run_kw)
+
+        try:
+            return self.middleware.call_sync(
+                'update.install_scale',
+                mounted,
+                lambda progress, description: job.set_progress((0.5 + 0.5 * progress) * 100, description),
+            )
+        finally:
+            shutil.rmtree(mounted)
